+++ ballast-debugger-idea-plugin/src/main/kotlin/com/copperleaf/ballast/debugger/ui/widgets/States.kt
@@ -1,9 +1,11 @@
 package com.copperleaf.ballast.debugger.ui.widgets
 
+import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.text.selection.SelectionContainer
 import androidx.compose.material.ListItem
+import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -41,6 +43,14 @@ fun StateSnapshotSummary(
 ) {
     ListItem(
         modifier = Modifier
+            .onHoverState { Modifier.highlight() }
+            .then(
+                if (uiState.focusedDebuggerEventUuid == stateSnapshot.uuid) {
+                    Modifier.background(MaterialTheme.colors.onSurface.copy(alpha = 0.1f))
+                } else {
+                    Modifier
+                }
+            )
             .clickable {
                 postInput(
                     DebuggerContract.Inputs.FocusEvent(