+++ ballast-test/src/commonMain/kotlin/com/copperleaf/ballast/test/internal/vm/TestEventHandler.kt
@@ -2,9 +2,7 @@ package com.copperleaf.ballast.test.internal.vm
 
 import com.copperleaf.ballast.EventHandler
 import com.copperleaf.ballast.EventHandlerScope
-import kotlinx.coroutines.ExperimentalCoroutinesApi
 
-@ExperimentalCoroutinesApi
 internal class TestEventHandler<Inputs : Any, Events : Any, State : Any>(
     private val eventHandlerDelegate: EventHandler<Inputs, Events, State>
 ) : EventHandler<TestViewModel.Inputs<Inputs>, Events, State> {