+++ ballast-test/src/commonMain/kotlin/com/copperleaf/ballast/test/internal/vm/TestSideEffectScope.kt
@@ -3,9 +3,7 @@ package com.copperleaf.ballast.test.internal.vm
 import com.copperleaf.ballast.SideEffectScope
 import kotlinx.coroutines.CompletableDeferred
 import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.ExperimentalCoroutinesApi
 
-@ExperimentalCoroutinesApi
 internal class TestSideEffectScope<Inputs : Any, Events : Any, State : Any>(
     private val sideEffectScopeDelegate: SideEffectScope<TestViewModel.Inputs<Inputs>, Events, State>,
 ) : SideEffectScope<Inputs, Events, State>, CoroutineScope by sideEffectScopeDelegate {