+++ ballast-test/src/commonMain/kotlin/com/copperleaf/ballast/test/internal/vm/ViewModelWrapper.kt
@@ -8,7 +8,6 @@ import kotlinx.coroutines.channels.SendChannel
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.selects.SelectClause2
 
-@ExperimentalCoroutinesApi
 internal class ViewModelWrapper<Inputs : Any, Events : Any, State : Any>(
     private val delegate: BallastViewModel<TestViewModel.Inputs<Inputs>, Events, State>,
 ) : BallastViewModel<Inputs, Events, State> {
@@ -16,6 +15,7 @@ internal class ViewModelWrapper<Inputs : Any, Events : Any, State : Any>(
     override val name: String
         get() = delegate.name
 
+    @ExperimentalCoroutinesApi
     override val isClosedForSend: Boolean
         get() = delegate.isClosedForSend
 
@@ -26,6 +26,7 @@ internal class ViewModelWrapper<Inputs : Any, Events : Any, State : Any>(
         return delegate.close(cause)
     }
 
+    @ExperimentalCoroutinesApi
     override fun invokeOnClose(handler: (cause: Throwable?) -> Unit) {
         delegate.invokeOnClose(handler)
     }